requirements.txt
openai-agents>=0.4.1
pinecone>=3.0.0
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
feedparser>=6.0.10
lxml>=4.9.3

